requirements.txt
pyautogen==0.1.14
streamlit==1.31.0
openai==0.28.1
requests==2.31.0
beautifulsoup4==4.12.3
duckduckgo-search==4.2.0
langchain==0.1.7
langchain-community==0.0.20
newspaper3k==0.2.8
lxml_html_clean==0.1.0
lxml==4.9.4
nltk==3.8.1 